Personal_AI_Employee/requirements.txt
google-auth-httplib2
google-auth-oauthlib
google-api-python-client
google-auth
python-dotenv
watchdog
plyer
playwright